Add files via upload
Split the frontend and tracking function up initially. Frontend only prints the selected bike as a placeholder for now and the tracking function does not put the data into the correct row. I will continue to work on the frontend and try to make it so that it queries the database depending on the bike selected and output the corresponding row from the database. Tracking function would most likely have to be merged with the bike files/publishers